+++ NEWS.md
@@ -1,3 +1,6 @@
-# geometr 0.0.0.9000
+geometr 0.0.0.9000
+=========================
+### IMPROVEMENTS
+* Working beta `geomet_daily_mean` function with station and date parameters
+* Tests added for `geomet_daily_mean`
 
-* Added a `NEWS.md` file to track changes to the package.